refactor: Move modal and input selects into `Select` folder
