+++ src/components/Select/Item/styled.tsx
@@ -1,8 +1,8 @@
 import styled from 'styled-components';
 import { em } from 'polished';
 
-import { ListItem } from '../../../ListItem';
-import { boxSizing } from '../../../../modules/box-sizing';
+import { ListItem } from '../../ListItem';
+import { boxSizing } from '../../../modules/box-sizing';
 
 export const StyledListItem = styled(ListItem)`
   ${boxSizing};
