+++ src/components/Select/ModalSelect/component.stories.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 
-import { Sections } from '../../modules/sections';
-import { Icon } from '../Icon';
+import { Sections } from '../../../modules/sections';
+import { Icon } from '../../Icon';
 
 // @ts-ignore
 import pic from './pic.png';
