+++ src/index.tsx
@@ -35,8 +35,8 @@ export { SolidAvatar } from './components/SolidAvatar';
 export { BuyWithMoonPay } from './components/BuyWithMoonPay';
 export { Space } from './components/Space';
 export { Modal } from './components/Modal';
-export { ModalSelect } from './components/ModalSelect';
+export { ModalSelect } from './components/Select/ModalSelect';
 export { ListItem } from './components/ListItem';
 export { Card } from './components/Card';
 export { Table } from './components/Table';
-export { InputSelect } from './components/InputSelect';
+export { InputSelect } from './components/Select/InputSelect';
